Code/FastAPIProject/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
pyjwt
psycopg2-binary
pydantic[email]
python-multipart
